+++ frontend/app/organization/team/properties/definitions/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { TeamMemberPropertyDefinitionsPanel } from "@/components/TeamMemberPropertyDefinitionsPanel";
+
+export default function TeamPropertyDefinitionsPage() {
+  return <TeamMemberPropertyDefinitionsPanel />;
+}